// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface JobsDocumentDataTagsItem {
  /**
   * Tag field in *job → Tags*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.tags[].tag
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  tag: prismic.KeyTextField;
}

/**
 * Content for job documents
 */
interface JobsDocumentData {
  /**
   * Title field in *job*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  title: prismic.KeyTextField;

  /**
   * Company field in *job*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.company
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  company: prismic.KeyTextField;

  /**
   * Location field in *job*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.location
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  location: prismic.KeyTextField;

  /**
   *  Contract Type field in *job*
   *
   * - **Field Type**: Select
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.contract_type
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/select
   */
  contract_type: prismic.SelectField<
    "CDD" | "CDI" | "Freelance" | "Stage" | "Alternance"
  >;

  /**
   * Salary field in *job*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.salary
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  salary: prismic.KeyTextField;

  /**
   * Tags field in *job*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.tags[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  tags: prismic.GroupField<Simplify<JobsDocumentDataTagsItem>>;

  /**
   * Short Description field in *job*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.short_description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  short_description: prismic.KeyTextField;

  /**
   * Description field in *job*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  description: prismic.RichTextField;

  /**
   * Published Date field in *job*
   *
   * - **Field Type**: Date
   * - **Placeholder**: *None*
   * - **API ID Path**: jobs.published_date
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/date
   */
  published_date: prismic.DateField;
}

/**
 * job document from Prismic
 *
 * - **API ID**: `jobs`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type JobsDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<Simplify<JobsDocumentData>, "jobs", Lang>;